separate components to cild components and parent components. added general components for table, lists, etc. added user page and rediraction from user in a table. fixed landing page
