+++ Client/src/assets/tableStyle.css
@@ -1,0 +1,13 @@
+.q-table th {
+    font-weight: bold;
+      color: #012d5d;
+
+
+}
+
+.table_header {
+  font-weight: bold;
+    color: #012d5d;
+
+
+}
