+++ Client/src/i18n/en-US/index.js
@@ -54,7 +54,7 @@ export default {
     location:'at'
   }  ,
   groups:{
-    name:'Group \'%d\' \'%s\''    //make number and description from api
+    name:'Group '    //make number and description from api
   },
   group:{
     groupDetails:'Group details',
@@ -86,6 +86,9 @@ export default {
     posticks:'New notes',
     markAsRead:'Mark as read',
     post:'Post note'
+ },
+ user:{
+    userPageHeader:'User details'
  } 
 }
