+++ Client/src/pages/admin/AllGroups.vue
@@ -1,36 +1,18 @@
 <template>
-<q-page>
-  <div class="q-pa-md" style="max-width: 355px">
-    <h3 class="group_header  q-mb-none">{{$t('app.groups')}}</h3>
-    <q-list bordered separator class="list-items">
-      <q-item to="/groups/id" clickable v-ripple>
-        <q-item-section>{{$t('groups.name')}}</q-item-section>
-      </q-item>
-
-      <q-item to="/groups/id" clickable v-ripple>
-        <q-item-section>
-          <q-item-label>{{$t('groups.name')}}</q-item-label>
-        </q-item-section>
-      </q-item>
-
-      <q-item to="/groups/id" clickable v-ripple>
-        <q-item-section>
-          <q-item-label>{{$t('groups.name')}}</q-item-label>
-        </q-item-section>
-      </q-item>
-    </q-list>
-  </div>
-  </q-page>
+<groups-list></groups-list>
 </template>
-<style scoped lang="scss">
-.group_header,
-.q-table th {
-  font-weight: bold;
-}
 
-.list-items{
-    font-weight: bold;
-  font-size: large;
+<script>
+import groupsList from 'components/list/groupsList.vue'
+
+export default {
+  components: { groupsList }
 }
+</script>
+
+
+
+<style scoped lang="scss">
+@import "assets/groupStyle.css"
 </style>
