+++ Client/src/pages/admin/mockdata.js
@@ -1,0 +1,102 @@
+export const mockRows = [
+  {
+    id: 1,
+    name: "תומר פיזון",
+    phone: "0526831999",
+    group: 1,
+    area: "תל אביב",
+    message: true,
+  },
+  {
+    id: 2,
+    name: "מאיה קימל",
+    phone: "054969614",
+    group: 3,
+    area: "הרצליה",
+    message: true,
+
+  },
+  {
+    id: 3,
+    name: "ברק עופר",
+    phone: "0526834333",
+    group: 1,
+    area: "תל אביב",
+  },
+  {
+    id: 4,
+        name: "נינט לוי",
+    phone: "0526831999",
+    group: 1,
+    area: "תל אביב",
+    message: true,
+  },
+  {
+    id: 5,
+        name: "איתי שרון",
+    phone: "0526544599",
+    group: 1,
+    area: "תל אביב",
+  },
+  {
+        name: "עומר אדם",
+    phone: "0526831999",
+    group: 2,
+    area: "רמת גן",
+  },
+  {
+    id: 7,
+    name: "אליהו ענבי",
+    phone: "0527766455",
+    group: 3,
+    area: "הרצליה",
+  },
+  // {
+  //   id: 8,
+  //   name: "דני דנקר",
+  //   phone: "0526831455",
+  //   group: 3,
+  //   area: "הרצליה",
+  // },
+  // {
+  //   id: 9,
+  //   name: "קרן פידרמן",
+  //   phone: "054696989",
+  //   group: 1,
+  //   area: "תל אביב",
+  // },
+  // {
+  //   id: 10,
+  //   name: "אופק גלר",
+  //   phone: "0521211999",
+  //   group: 2,
+  //   area: "רמת גן",
+  // }
+]  
+export const mockTrainers = [
+  {
+    id: 1,
+    name: "דנה טושינקי",
+    phone: "0526831999",
+    groups: "1"
+  },
+  {
+    id: 2,
+    name: "גל לימנואל",
+    phone: "0549691614",
+    groups: "2",
+  },
+  {
+    id: 3,
+    name: "בועז יעקב",
+    phone: "0526834333",
+    groups: "3",
+  },
+  {
+    id: 4,
+    name: "דניאל רוט",
+    phone: "0526831999",
+    groups: "2",
+  }
+];
+
